+++ rollup.config.js
@@ -1,4 +1,6 @@
 import babel from "rollup-plugin-babel";
+import resolve from "rollup-plugin-node-resolve";
+import commonjs from "rollup-plugin-commonjs";
 
 export default {
     input: "./src/index.js",
@@ -8,6 +10,10 @@ export default {
         format: "umd"
     },
     plugins: [
+        resolve(),
+        commonjs({
+            include: "node_modules/**"
+        }),
         babel({
             exclude: "node_modules/**"
         })