+++ src/reusables.js
@@ -1,5 +1,3 @@
-import "@babel/polyfill";
-
 export function asArray(x) {
     if (!x) return [];
     return Array.isArray(x) ? x : [x];