+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 -   Replace webpack by rollup (easier to configure).
 -   Drop webpack completely
 -   Drop support for es3. You can add it by following the guide.
+-   Correct a warning about fs module for front-end builds.
+-   Remove a circle dependency (dataframe.js -> groupedDataframe.js -> dataframe.js)
 
 ---